Refactor: move show_translation_steps to a separate module
- Moved show_translation_steps and export_steps to show_steps.py
- Cleaned up import structure for better maintainability
